Revise Gaussian Elimination 第一步驟 Search for maximum in this column 是要比較絕對值

+++ Project1/Project1/Matrix.h
@@ -39,7 +39,7 @@ public:
 	/*
 	 * Rank
 	 */
-	double rank();
+	unsigned int rank();
 
 	Matrix trans();
 	double det();
